+++ Core/Component/PhysxComponent.h
@@ -28,7 +28,7 @@ public:
 	virtual ~CPhysxComponent();
 
 	virtual void PhysxUpdate();
-	virtual void FixedUpdate(float ElapsedTime);
+	virtual void FixedUpdate(double ElapsedTime);
 
 	virtual void Destroy();
